+++ R/MazamaSpatialUtils.R
@@ -181,11 +181,11 @@ setSpatialDataDir <- function(dataDir) {
 
 #' @keywords conversion
 #' @export
-#' @title Convert Between ISO2 and ISO3 Country Codes
+#' @title Convert From ISO2 to ISO3 Country Codes
 #' @param countryCodes vector of country codes to be converted
-#' @description Converts a vector of ISO 3166-1 alpha-2 codes to the corresponding ISO 3166-1 alpha-3 codes or vice versa.
-#' @return A vector of ISO country codes
-codeToCode <- function(countryCodes) {
+#' @description Converts a vector of ISO 3166-1 alpha-2 codes to the corresponding ISO 3166-1 alpha-3 codes.
+#' @return A vector of ISO3 country codes
+iso2ToIso3 <- function(countryCodes) {
   countryTable <- MazamaSpatialUtils::SimpleCountries@data
   nonMissingCountryCodes <- countryCodes[!is.na(countryCodes)]
   if ( all(stringr::str_length(nonMissingCountryCodes) == 2) ) {
@@ -193,13 +193,27 @@ codeToCode <- function(countryCodes) {
     allISO3 <- countryTable$ISO3
     names(allISO3) <- countryTable$countryCode
     return(as.character(allISO3[countryCodes]))
-  } else if ( all(stringr::str_length(nonMissingCountryCodes) == 3) ) {
+  } else {
+    stop('countryCodes must be all ISO 3166-1 alpha-2', call.=FALSE)
+  }
+}
+
+#' @keywords conversion
+#' @export
+#' @title Convert Betweenrom ISO3 to ISO2 Country Codes
+#' @param countryCodes vector of country codes to be converted
+#' @description Converts a vector of ISO 3166-1 alpha-3 codes to the corresponding ISO 3166-1 alpha-2 codes.
+#' @return A vector of ISO2 country codes
+iso3ToIso2 <- function(countryCodes) {
+  countryTable <- MazamaSpatialUtils::SimpleCountries@data
+  nonMissingCountryCodes <- countryCodes[!is.na(countryCodes)]
+  if ( all(stringr::str_length(nonMissingCountryCodes) == 3) ) {
     # Create a vector of ISO2 identified by ISO3
     allISO2 <- countryTable$countryCode
     names(allISO2) <- countryTable$ISO3
     return(as.character(allISO2[countryCodes]))    
   } else {
-    stop('countryCodes must be either all ISO 3166-1 alpha-2 or all ISO 3166-1 alpha-3', call.=FALSE)
+    stop('countryCodes must be all ISO 3166-1 alpha-3', call.=FALSE)
   }
 }